refact: split functions for better redability and manutenability
