+++ scripts/utils/delay.ts
@@ -1,0 +1,4 @@
+// Simple async delay function
+export function delay(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
